Corrected `poly` expression in .truncpoi (line 62)
The expression for computing the polynomial (assigned to the`poly` variable on line 62) was originally written to span multiple lines, just as it appeared in the SAS code on which this R package was based. However, in R, the portion of the expression (the g1^2 term) that was relegated to a second line of code is ignored. The effect on the resulting confidence intervals is very small and often not noticeable (usually less than 1%), but could nonetheless be significant depending on the values used in the input vector. This version corrects the expression by having it all written on a single line prior to its assignment to the `poly` variable.